apps/backend-legacy/requirements.txt
# requirements.txt
# Project dependencies with exact version locking
Django==5.2.4
djangorestframework==3.16.0
channels==4.3.0
channels-redis==4.2.1
Pillow==11.2.1
requests==2.32.4
redis==6.2.0
python-dotenv==1.1.1
pip==25.1.1
mysqlclient==2.2.7
